requirements.txt
modal-client>=0.54.0
gradio>=3.50.0
transformers>=4.15.0
torch>=1.9.0
tensorflow>=2.5.0
datasets>=1.11.0
huggingface-hub>=0.4.0
wandb>=0.15.0
